+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## WIP
 
 - Update content dashboard widget to sort models by latest updated item
+- Fix login dialog loading in spaces
 
 ## 2.11.1 (2025-02-25)